+++ mocks/attention-store.go
@@ -9,7 +9,7 @@ import (
 )
 
 type AttentionStore struct {
-	AnswerStub        func(context.Context, pkg.ItemID, string, string, pkg.Decision) (*pkg.Item, error)
+	AnswerStub        func(context.Context, pkg.ItemID, string, string, pkg.Decision, *pkg.Answer) (*pkg.Item, error)
 	answerMutex       sync.RWMutex
 	answerArgsForCall []struct {
 		arg1 context.Context
@@ -17,6 +17,7 @@ type AttentionStore struct {
 		arg3 string
 		arg4 string
 		arg5 pkg.Decision
+		arg6 *pkg.Answer
 	}
 	answerReturns struct {
 		result1 *pkg.Item
@@ -113,7 +114,7 @@ type AttentionStore struct {
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *AttentionStore) Answer(arg1 context.Context, arg2 pkg.ItemID, arg3 string, arg4 string, arg5 pkg.Decision) (*pkg.Item, error) {
+func (fake *AttentionStore) Answer(arg1 context.Context, arg2 pkg.ItemID, arg3 string, arg4 string, arg5 pkg.Decision, arg6 *pkg.Answer) (*pkg.Item, error) {
 	fake.answerMutex.Lock()
 	ret, specificReturn := fake.answerReturnsOnCall[len(fake.answerArgsForCall)]
 	fake.answerArgsForCall = append(fake.answerArgsForCall, struct {
@@ -122,13 +123,14 @@ func (fake *AttentionStore) Answer(arg1 context.Context, arg2 pkg.ItemID, arg3 s
 		arg3 string
 		arg4 string
 		arg5 pkg.Decision
-	}{arg1, arg2, arg3, arg4, arg5})
+		arg6 *pkg.Answer
+	}{arg1, arg2, arg3, arg4, arg5, arg6})
 	stub := fake.AnswerStub
 	fakeReturns := fake.answerReturns
-	fake.recordInvocation("Answer", []interface{}{arg1, arg2, arg3, arg4, arg5})
+	fake.recordInvocation("Answer", []interface{}{arg1, arg2, arg3, arg4, arg5, arg6})
 	fake.answerMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3, arg4, arg5)
+		return stub(arg1, arg2, arg3, arg4, arg5, arg6)
 	}
 	if specificReturn {
 		return ret.result1, ret.result2
@@ -142,17 +144,17 @@ func (fake *AttentionStore) AnswerCallCount() int {
 	return len(fake.answerArgsForCall)
 }
 
-func (fake *AttentionStore) AnswerCalls(stub func(context.Context, pkg.ItemID, string, string, pkg.Decision) (*pkg.Item, error)) {
+func (fake *AttentionStore) AnswerCalls(stub func(context.Context, pkg.ItemID, string, string, pkg.Decision, *pkg.Answer) (*pkg.Item, error)) {
 	fake.answerMutex.Lock()
 	defer fake.answerMutex.Unlock()
 	fake.AnswerStub = stub
 }
 
-func (fake *AttentionStore) AnswerArgsForCall(i int) (context.Context, pkg.ItemID, string, string, pkg.Decision) {
+func (fake *AttentionStore) AnswerArgsForCall(i int) (context.Context, pkg.ItemID, string, string, pkg.Decision, *pkg.Answer) {
 	fake.answerMutex.RLock()
 	defer fake.answerMutex.RUnlock()
 	argsForCall := fake.answerArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5
+	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5, argsForCall.arg6
 }
 
 func (fake *AttentionStore) AnswerReturns(result1 *pkg.Item, result2 error) {